Move sub clauses to mixins
