+++ packages/graphql/src/translate/cypher-builder/clauses/Match.ts
@@ -19,21 +19,15 @@
 
 import type { CypherEnvironment } from "../Environment";
 import { MatchableElement, MatchParams, Pattern } from "../Pattern";
-import { Where, WhereParams } from "../sub-clauses/Where";
 import { Clause } from "./Clause";
-import { Return, ReturnColumn } from "./Return";
-import { Variable } from "../variables/Variable";
-import { ComparisonOp, eq } from "../operations/comparison";
-import { PropertyRef } from "../PropertyRef";
-import { and, BooleanOp } from "../operations/boolean";
-import { SetClause, SetParam } from "../sub-clauses/Set";
 import { compileCypherIfExists } from "../utils";
+import { WithReturn } from "./mixins/WithReturn";
+import { applyMixins } from "./utils/apply-mixin";
+import { WithWhere } from "./mixins/WithWhere";
+import { WithSet } from "./mixins/WIthSet";
 
 export class Match<T extends MatchableElement = any> extends Clause {
     private pattern: Pattern<T>;
-    private whereSubClause: Where | undefined;
-    private returnStatement: Return | undefined;
-    private setSubClause: SetClause | undefined;
 
     constructor(variable: T | Pattern<T>, parameters: MatchParams<T> = {}, parent?: Clause) {
         super(parent);
@@ -45,54 +39,6 @@ export class Match<T extends MatchableElement = any> extends Clause {
         this.addChildren(this.pattern);
     }
 
-    public where(input: WhereParams): this;
-    public where(target: Variable, params: Record<string, Variable>): this;
-    public where(input: WhereParams | Variable, params?: Record<string, Variable>): this {
-        const whereInput = this.createWhereInput(input, params);
-        if (!whereInput) return this;
-
-        if (!this.whereSubClause) {
-            const whereClause = new Where(this, whereInput);
-            this.whereSubClause = whereClause;
-        } else {
-            this.and(whereInput);
-        }
-        return this;
-    }
-
-    public and(input: WhereParams): this;
-    public and(target: Variable, params: Record<string, Variable>): this;
-    public and(input: WhereParams | Variable, params?: Record<string, Variable>): this {
-        if (!this.whereSubClause) throw new Error("Cannot and without a where");
-        const whereInput = this.createWhereInput(input, params);
-        if (whereInput) {
-            this.whereSubClause.and(whereInput);
-        }
-        return this;
-    }
-
-    public set(...params: SetParam[]): this {
-        if (!this.setSubClause) {
-            this.setSubClause = new SetClause(this, params);
-        } else {
-            this.setSubClause.addParams(...params);
-        }
-        return this;
-    }
-
-    public return(...columns: ReturnColumn[]): Return;
-    public return(starOrColumn: "*" | ReturnColumn, ...columns: ReturnColumn[]): Return;
-    public return(starOrColumn: "*" | ReturnColumn | undefined, ...columns: ReturnColumn[]): Return {
-        if (this.returnStatement) throw new Error("Cannot set multiple return statements in Match clause");
-        if (!starOrColumn) {
-            this.returnStatement = new Return();
-        } else {
-            this.returnStatement = new Return(starOrColumn, ...columns);
-        }
-        this.addChildren(this.returnStatement);
-        return this.returnStatement;
-    }
-
     public getCypher(env: CypherEnvironment): string {
         const nodeCypher = this.pattern.getCypher(env);
 
@@ -102,32 +48,7 @@ export class Match<T extends MatchableElement = any> extends Clause {
 
         return `MATCH ${nodeCypher}${whereCypher}${setCypher}${returnCypher}`;
     }
-
-    private createWhereInput(
-        input: WhereParams | Variable,
-        params: Record<string, Variable> | undefined
-    ): WhereParams | undefined {
-        if (input instanceof Variable) {
-            const generatedOp = this.variableAndObjectToOperation(input, params || {});
-            return generatedOp;
-        }
-        return input;
-    }
-
-    /** Transforms a simple input into an operation sub tree */
-    private variableAndObjectToOperation(
-        target: Variable,
-        params: Record<string, Variable>
-    ): BooleanOp | ComparisonOp | undefined {
-        let operation: BooleanOp | ComparisonOp | undefined;
-        for (const [key, value] of Object.entries(params)) {
-            const property = new PropertyRef(target, key);
-            const eqOp = eq(property, value);
-            if (!operation) operation = eqOp;
-            else {
-                operation = and(operation, eqOp);
-            }
-        }
-        return operation;
-    }
 }
+
+export interface Match extends WithReturn, WithWhere, WithSet {}
+applyMixins(Match, [WithReturn, WithWhere, WithSet]);
