+++ packages/graphql/src/translate/cypher-builder/clauses/mixins/WIthSet.ts
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) "Neo4j"
+ * Neo4j Sweden AB [http://neo4j.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { SetClause, SetParam } from "../../sub-clauses/Set";
+import { ClauseMixin } from "./ClauseMixin";
+
+export abstract class WithSet extends ClauseMixin {
+    protected setSubClause: SetClause | undefined;
+
+    public set(...params: SetParam[]): this {
+        if (!this.setSubClause) {
+            this.setSubClause = new SetClause(this, params);
+        } else {
+            this.setSubClause.addParams(...params);
+        }
+        return this;
+    }
+}
